requirements.txt
voyageai
langchain_community
anthropic
scikit-learn
numpy
matplotlib
openai
python-dotenv
flask
flask_session
flask_sqlalchemy
